gpt4.1-deepgram-elevenlabs-vapi/pyproject.toml
[project]
name = "gpt4.1-deepgram-elevenlabs-vapi"
version = "0.1.0"
description = "Voice agent using Vapi orchestrator with GPT-4.1, Deepgram STT, ElevenLabs TTS, and Plivo telephony"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "vapi-server-sdk>=0.5.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "plivo>=4.59.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "phonenumbers>=8.13.0",
    "httpx>=0.27.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "pre-commit>=4.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "faster-whisper>=1.2.1",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP", "SIM", "RUF"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
